requirements.txt
langchain-community
langchain-classic
langchain-text-splitters
langchain-huggingface
chromadb
langchain-chroma
langchain-groq
sentence-transformers
unstructured
unstructured[pdf]
nltk
streamlit
python-dotenv